+++ Core/Source/Rendering/RenderCommands.h
@@ -26,6 +26,6 @@ namespace Engine {
 
 		static void RenderMesh(const Mesh& mesh, const MaterialInstance& material);
 		static void RenderMesh(const VertexArrayObject& mesh, const Shader& shader);
-		static void RenderPoints(const VertexArrayObject& vertexArray, uint64_t count, const Shader& shader);
+		static void RenderPoints(const VertexArrayObject& vertexArray, uint32_t count, const Shader& shader);
 	};
 }
